test: move hand tests from main.rs to hand.rs
