+++ src/hand.rs
@@ -861,3 +861,1517 @@ impl Hand {
         false
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::Hand;
+    use crate::hand::error::HandErr;
+
+    #[test]
+    fn yaku_kokushi() {
+        let out = Hand::new(
+            vec![
+                "1s".to_string(),
+                "9s".to_string(),
+                "1m".to_string(),
+                "9m".to_string(),
+                "1p".to_string(),
+                "9p".to_string(),
+                "Ew".to_string(),
+                "Sw".to_string(),
+                "Ww".to_string(),
+                "Nw".to_string(),
+                "gd".to_string(),
+                "rd".to_string(),
+                "wwd".to_string(),
+            ],
+            "wd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_kokushi());
+        assert!(out.is_kokushi13sided());
+        let out = Hand::new(
+            vec![
+                "1s".to_string(),
+                "9s".to_string(),
+                "1m".to_string(),
+                "9m".to_string(),
+                "1p".to_string(),
+                "9p".to_string(),
+                "Ew".to_string(),
+                "Sw".to_string(),
+                "Ww".to_string(),
+                "Nw".to_string(),
+                "gd".to_string(),
+                "wwd".to_string(),
+                "rd".to_string(),
+            ],
+            "rd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_kokushi());
+        assert!(!out.is_kokushi13sided());
+        let out = Hand::new(
+            vec![
+                "9s".to_string(),
+                "9s".to_string(),
+                "1m".to_string(),
+                "9m".to_string(),
+                "1p".to_string(),
+                "9p".to_string(),
+                "Ew".to_string(),
+                "Sw".to_string(),
+                "Ww".to_string(),
+                "Nw".to_string(),
+                "gd".to_string(),
+                "rd".to_string(),
+                "wwd".to_string(),
+            ],
+            "wd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_kokushi());
+    }
+
+    #[test]
+    fn yaku_daisuushii() {
+        let out = Hand::new(
+            vec![
+                "EEEEw".to_string(),
+                "SSSw".to_string(),
+                "WWWw".to_string(),
+                "NNNw".to_string(),
+                "99s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_daisuushii());
+        let out = Hand::new(
+            vec![
+                "EEEEw".to_string(),
+                "SSw".to_string(),
+                "WWWw".to_string(),
+                "NNNw".to_string(),
+                "999s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_daisuushii());
+    }
+
+    #[test]
+    fn yaku_shousuushi() {
+        let out = Hand::new(
+            vec![
+                "EEEEw".to_string(),
+                "SSw".to_string(),
+                "WWWw".to_string(),
+                "NNNw".to_string(),
+                "999s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_shousuushii());
+        let out = Hand::new(
+            vec![
+                "EEEEw".to_string(),
+                "22s".to_string(),
+                "WWWw".to_string(),
+                "NNNw".to_string(),
+                "999s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_shousuushii());
+    }
+
+    #[test]
+    fn yaku_suukantsu() {
+        let out = Hand::new(
+            vec![
+                "EEEEw".to_string(),
+                "2222p".to_string(),
+                "1111mo".to_string(),
+                "7777s".to_string(),
+                "99s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_suukantsu());
+        let out = Hand::new(
+            vec![
+                "EEEw".to_string(),
+                "2222p".to_string(),
+                "1111mo".to_string(),
+                "7777s".to_string(),
+                "99s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_suukantsu());
+    }
+
+    #[test]
+    fn yaku_daichiishin() {
+        let out = Hand::new(
+            vec![
+                "SSw".to_string(),
+                "rrd".to_string(),
+                "ggd".to_string(),
+                "wwd".to_string(),
+                "NNw".to_string(),
+                "SSw".to_string(),
+                "EEw".to_string(),
+            ],
+            "Ew".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_daichiishin());
+        let out = Hand::new(
+            vec![
+                "WWw".to_string(),
+                "NNNw".to_string(),
+                "gggd".to_string(),
+                "rrrd".to_string(),
+                "EEEw".to_string(),
+            ],
+            "Ew".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_daichiishin());
+    }
+
+    #[test]
+    fn yaku_tsuuiisou() {
+        let out = Hand::new(
+            vec![
+                "SSw".to_string(),
+                "NNNw".to_string(),
+                "gggd".to_string(),
+                "rrrd".to_string(),
+                "EEEw".to_string(),
+            ],
+            "Ew".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_tsuuiisou());
+        let out = Hand::new(
+            vec![
+                "11s".to_string(),
+                "NNNw".to_string(),
+                "gggd".to_string(),
+                "rrrd".to_string(),
+                "EEEw".to_string(),
+            ],
+            "Ew".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_tsuuiisou());
+    }
+
+    #[test]
+    fn yaku_cuurenpoutou() {
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "234s".to_string(),
+                "55s".to_string(),
+                "678s".to_string(),
+                "999s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_chuurenpoutou());
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "234s".to_string(),
+                "55m".to_string(),
+                "678s".to_string(),
+                "999s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chuurenpoutou());
+        let out = Hand::new(
+            vec![
+                "123s".to_string(),
+                "234s".to_string(),
+                "555s".to_string(),
+                "678s".to_string(),
+                "99s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chuurenpoutou());
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "234s".to_string(),
+                "678s".to_string(),
+                "999s".to_string(),
+                "55s".to_string(),
+            ],
+            "5s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_chuurenpoutou9sided());
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "234s".to_string(),
+                "678s".to_string(),
+                "55s".to_string(),
+                "999s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chuurenpoutou9sided());
+    }
+
+    #[test]
+    fn yaku_ryuuiisou() {
+        let out = Hand::new(
+            vec![
+                "234s".to_string(),
+                "234s".to_string(),
+                "66s".to_string(),
+                "gggd".to_string(),
+                "888s".to_string(),
+            ],
+            "8s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_ryuuiisou());
+        let out = Hand::new(
+            vec![
+                "345s".to_string(),
+                "234s".to_string(),
+                "66s".to_string(),
+                "gggd".to_string(),
+                "888s".to_string(),
+            ],
+            "8s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_ryuuiisou());
+
+        let out = Hand::new(
+            vec![
+                "234s".to_string(),
+                "234s".to_string(),
+                "666s".to_string(),
+                "gggd".to_string(),
+                "99s".to_string(),
+            ],
+            "9s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_ryuuiisou());
+    }
+
+    #[test]
+    fn yaku_chinroutou() {
+        let out = Hand::new(
+            vec![
+                "111so".to_string(),
+                "1111m".to_string(),
+                "999s".to_string(),
+                "999p".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_chinroutou());
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "1111m".to_string(),
+                "999s".to_string(),
+                "999p".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chinroutou());
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "1111m".to_string(),
+                "789s".to_string(),
+                "999p".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chinroutou());
+    }
+
+    #[test]
+    fn yaku_suuankoutankiwait() {
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "888p".to_string(),
+                "777s".to_string(),
+                "111s".to_string(),
+                "11m".to_string(),
+            ],
+            "1m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_suuankoutankiwait());
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "888p".to_string(),
+                "777s".to_string(),
+                "11m".to_string(),
+                "111s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_suuankoutankiwait());
+    }
+
+    #[test]
+    fn yaku_suuankou() {
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "888p".to_string(),
+                "777s".to_string(),
+                "111s".to_string(),
+                "11m".to_string(),
+            ],
+            "1m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_suuankou(false));
+
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "888p".to_string(),
+                "777s".to_string(),
+                "11m".to_string(),
+                "111s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_suuankou(true));
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "888p".to_string(),
+                "777s".to_string(),
+                "11m".to_string(),
+                "111s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_suuankou(false));
+    }
+
+    #[test]
+    fn yaku_daisangen() {
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "gggd".to_string(),
+                "wwwwd".to_string(),
+                "88p".to_string(),
+                "567p".to_string(),
+            ],
+            "6p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_daisangen());
+        let out = Hand::new(
+            vec![
+                "rrrrd".to_string(),
+                "ggggd".to_string(),
+                "wwwwd".to_string(),
+                "88p".to_string(),
+                "567p".to_string(),
+            ],
+            "6p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_daisangen());
+
+        let out = Hand::new(
+            vec![
+                "rrrrd".to_string(),
+                "gggd".to_string(),
+                "wwd".to_string(),
+                "888p".to_string(),
+                "567p".to_string(),
+            ],
+            "6p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_daisangen());
+    }
+
+    #[test]
+    fn yaku_chinitsu() {
+        let out = Hand::new(
+            vec![
+                "222p".to_string(),
+                "123p".to_string(),
+                "345p".to_string(),
+                "88p".to_string(),
+                "567p".to_string(),
+            ],
+            "6p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_chinitsu());
+        let out = Hand::new(
+            vec![
+                "222p".to_string(),
+                "123p".to_string(),
+                "345p".to_string(),
+                "88s".to_string(),
+                "567p".to_string(),
+            ],
+            "6p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chinitsu());
+    }
+
+    #[test]
+    fn yaku_sanshokudoukou() {
+        let out = Hand::new(
+            vec![
+                "222p".to_string(),
+                "222m".to_string(),
+                "222s".to_string(),
+                "11s".to_string(),
+                "456m".to_string(),
+            ],
+            "6m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_sanshokudoukou());
+        let out = Hand::new(
+            vec![
+                "222p".to_string(),
+                "2222m".to_string(),
+                "222s".to_string(),
+                "3333s".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_sanshokudoukou());
+        let out = Hand::new(
+            vec![
+                "222p".to_string(),
+                "333m".to_string(),
+                "222s".to_string(),
+                "11s".to_string(),
+                "333s".to_string(),
+            ],
+            "3s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_sanshokudoukou());
+    }
+
+    #[test]
+    fn yaku_pinfu() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "678p".to_string(),
+                "345s".to_string(),
+                "11s".to_string(),
+                "456m".to_string(),
+            ],
+            "6m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_pinfu());
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "678po".to_string(),
+                "345s".to_string(),
+                "11s".to_string(),
+                "456m".to_string(),
+            ],
+            "5m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_pinfu());
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "678p".to_string(),
+                "345s".to_string(),
+                "11s".to_string(),
+                "456m".to_string(),
+            ],
+            "5m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_pinfu());
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "678p".to_string(),
+                "345s".to_string(),
+                "11s".to_string(),
+                "rrrd".to_string(),
+            ],
+            "rd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_pinfu());
+    }
+
+    #[test]
+    fn yaku_menzentsumo() {
+        let out = Hand::new(
+            vec![
+                "11s".to_string(),
+                "22p".to_string(),
+                "33p".to_string(),
+                "44p".to_string(),
+                "55p".to_string(),
+                "66p".to_string(),
+                "77p".to_string(),
+            ],
+            "7p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_menzentsumo(true));
+
+        let out = Hand::new(
+            vec![
+                "11s".to_string(),
+                "22p".to_string(),
+                "33p".to_string(),
+                "44p".to_string(),
+                "55p".to_string(),
+                "66p".to_string(),
+                "77p".to_string(),
+            ],
+            "7p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_menzentsumo(false));
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "999p".to_string(),
+                "111p".to_string(),
+                "rrrdo".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_menzentsumo(false));
+    }
+
+    #[test]
+    fn yaku_chiitoitsu() {
+        let out = Hand::new(
+            vec![
+                "11s".to_string(),
+                "22p".to_string(),
+                "33p".to_string(),
+                "44p".to_string(),
+                "55p".to_string(),
+                "66p".to_string(),
+                "77p".to_string(),
+            ],
+            "7p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_chiitoitsu());
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "999p".to_string(),
+                "111p".to_string(),
+                "rrrdo".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chiitoitsu());
+    }
+
+    #[test]
+    fn yaku_chantaiyao() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "999p".to_string(),
+                "111p".to_string(),
+                "rrrdo".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_chantaiyao());
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "999p".to_string(),
+                "111p".to_string(),
+                "999m".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chantaiyao());
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "999p".to_string(),
+                "111p".to_string(),
+                "999m".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_chantaiyao());
+    }
+
+    #[test]
+    fn yaku_ittsuu() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "456p".to_string(),
+                "789p".to_string(),
+                "rrrdo".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_ittsuu());
+
+        let out = Hand::new(
+            vec![
+                "789m".to_string(),
+                "456m".to_string(),
+                "123m".to_string(),
+                "234m".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_ittsuu());
+        let out = Hand::new(
+            vec![
+                "123s".to_string(),
+                "789s".to_string(),
+                "456s".to_string(),
+                "234m".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_ittsuu());
+        let out = Hand::new(
+            vec![
+                "123m".to_string(),
+                "456m".to_string(),
+                "678m".to_string(),
+                "234m".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_ittsuu());
+    }
+
+    #[test]
+    fn yaku_sankantsu() {
+        let out = Hand::new(
+            vec![
+                "9999so".to_string(),
+                "123p".to_string(),
+                "SSSSw".to_string(),
+                "EEEEw".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_sankantsu());
+        let out = Hand::new(
+            vec![
+                "9999so".to_string(),
+                "123p".to_string(),
+                "SSSw".to_string(),
+                "EEEEw".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_sankantsu());
+    }
+
+    #[test]
+    fn yaku_honroutou() {
+        let out = Hand::new(
+            vec![
+                "999s".to_string(),
+                "111p".to_string(),
+                "SSSw".to_string(),
+                "EEEw".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_honroutou());
+
+        let out = Hand::new(
+            vec![
+                "999s".to_string(),
+                "123p".to_string(),
+                "SSSw".to_string(),
+                "EEEw".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_honroutou());
+
+        let out = Hand::new(
+            vec![
+                "999s".to_string(),
+                "111p".to_string(),
+                "111s".to_string(),
+                "999m".to_string(),
+                "99p".to_string(),
+            ],
+            "9p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_honroutou());
+    }
+
+    #[test]
+    fn yaku_junchantaiyao() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "123p".to_string(),
+                "999m".to_string(),
+                "789s".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_junchantaiyao());
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "123p".to_string(),
+                "999m".to_string(),
+                "789s".to_string(),
+                "ggd".to_string(),
+            ],
+            "gd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_junchantaiyao());
+
+        let out = Hand::new(
+            vec![
+                "111s".to_string(),
+                "111m".to_string(),
+                "999m".to_string(),
+                "999s".to_string(),
+                "11p".to_string(),
+            ],
+            "1p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_junchantaiyao());
+    }
+
+    #[test]
+    fn yaku_shousangen() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "wwwwd".to_string(),
+                "rrd".to_string(),
+                "gggd".to_string(),
+            ],
+            "gd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_shousangen());
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "wwwwd".to_string(),
+                "rrd".to_string(),
+                "234p".to_string(),
+            ],
+            "4p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_shousangen());
+    }
+
+    #[test]
+    fn yaku_honitsu() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "567p".to_string(),
+                "rrd".to_string(),
+                "gggd".to_string(),
+            ],
+            "gd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_honitsu());
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "567m".to_string(),
+                "rrd".to_string(),
+                "gggd".to_string(),
+            ],
+            "gd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_honitsu());
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "567p".to_string(),
+                "111p".to_string(),
+                "33p".to_string(),
+            ],
+            "3p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        //should be chinitsu not honitsu
+        assert!(!out.is_honitsu());
+    }
+
+    #[test]
+    fn yaku_sanankou() {
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "555m".to_string(),
+                "11s".to_string(),
+                "333p".to_string(),
+            ],
+            "3p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_sanankou(false));
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "222p".to_string(),
+                "555m".to_string(),
+                "11s".to_string(),
+                "333p".to_string(),
+            ],
+            "3p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_sanankou(true));
+
+        let out = Hand::new(
+            vec![
+                "123p".to_string(),
+                "123s".to_string(),
+                "555m".to_string(),
+                "11s".to_string(),
+                "333p".to_string(),
+            ],
+            "3p".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_sanankou(true));
+    }
+
+    #[test]
+    fn yaku_sanshokudoujun() {
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "234p".to_string(),
+                "234m".to_string(),
+                "234s".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_sanshokudoujun());
+
+        let out = Hand::new(
+            vec![
+                "678s".to_string(),
+                "234p".to_string(),
+                "234m".to_string(),
+                "234s".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_sanshokudoujun());
+
+        let out = Hand::new(
+            vec![
+                "111p".to_string(),
+                "234p".to_string(),
+                "234m".to_string(),
+                "rrrd".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_sanshokudoujun());
+    }
+
+    #[test]
+    fn yaku_toitoi() {
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "EEEw".to_string(),
+                "gggd".to_string(),
+                "222p".to_string(),
+                "11s".to_string(),
+            ],
+            "1s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_toitoi());
+
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "2222m".to_string(),
+                "EEEw".to_string(),
+                "11s".to_string(),
+                "gggd".to_string(),
+            ],
+            "gd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_toitoi());
+
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "2222m".to_string(),
+                "EEEw".to_string(),
+                "11s".to_string(),
+                "456so".to_string(),
+            ],
+            "5s".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_toitoi());
+    }
+
+    #[test]
+    fn yaku_yakuhai() {
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "234m".to_string(),
+                "22s".to_string(),
+                "234m".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert_eq!(out.is_yakuhai(), 1);
+        let out = Hand::new(
+            vec![
+                "EEEEwo".to_string(),
+                "234m".to_string(),
+                "22s".to_string(),
+                "234m".to_string(),
+                "rrrd".to_string(),
+            ],
+            "rd".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert_eq!(out.is_yakuhai(), 2);
+        let out = Hand::new(
+            vec![
+                "3333mo".to_string(),
+                "WWWm".to_string(),
+                "22s".to_string(),
+                "234m".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert_eq!(out.is_yakuhai(), 1);
+        let out = Hand::new(
+            vec![
+                "3333mo".to_string(),
+                "222m".to_string(),
+                "22s".to_string(),
+                "234m".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert_eq!(out.is_yakuhai(), 0);
+    }
+
+    #[test]
+    fn yaku_ryanpeikou() {
+        let out = Hand::new(
+            vec![
+                "123s".to_string(),
+                "123s".to_string(),
+                "789m".to_string(),
+                "789m".to_string(),
+                "77m".to_string(),
+            ],
+            "7m".to_string(),
+            "es".to_string(),
+            "ww".to_string(),
+        )
+        .unwrap();
+        //is open
+        assert!(out.is_ryanpeikou());
+
+        let out = Hand::new(
+            vec![
+                "123s".to_string(),
+                "123s".to_string(),
+                "789m".to_string(),
+                "678m".to_string(),
+                "77m".to_string(),
+            ],
+            "7m".to_string(),
+            "es".to_string(),
+            "ww".to_string(),
+        )
+        .unwrap();
+        //is open
+        assert!(!out.is_ryanpeikou());
+    }
+
+    #[test]
+    fn yaku_iipeko() {
+        let out = Hand::new(
+            vec![
+                "123s".to_string(),
+                "123s".to_string(),
+                "789m".to_string(),
+                "789m".to_string(),
+                "77m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        //is open
+        assert!(!out.is_iipeikou());
+
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "234m".to_string(),
+                "22s".to_string(),
+                "234m".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        //is open
+        assert!(!out.is_iipeikou());
+
+        let out = Hand::new(
+            vec![
+                "rrrd".to_string(),
+                "234m".to_string(),
+                "22s".to_string(),
+                "234m".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_iipeikou());
+    }
+
+    #[test]
+    fn yaku_tanyao() {
+        let out = Hand::new(
+            vec![
+                "rrrdo".to_string(),
+                "5555mo".to_string(),
+                "22s".to_string(),
+                "8888s".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_tanyao());
+        let out = Hand::new(
+            vec![
+                "333mo".to_string(),
+                "5555mo".to_string(),
+                "11s".to_string(),
+                "8888s".to_string(),
+                "678m".to_string(),
+            ],
+            "7m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(!out.is_tanyao());
+        let out = Hand::new(
+            vec![
+                "555mo".to_string(),
+                "678p".to_string(),
+                "22s".to_string(),
+                "333s".to_string(),
+                "345m".to_string(),
+            ],
+            "4m".to_string(),
+            "Es".to_string(),
+            "Ww".to_string(),
+        )
+        .unwrap();
+        assert!(out.is_tanyao());
+    }
+
+    #[test]
+    fn invalid_group_sequence_not_in_order() {
+        let out = Hand::new(
+            vec![
+                "135m".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "Sw".to_string(),
+            ],
+            "3s".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidGroup);
+    }
+
+    #[test]
+    fn invalid_group_size_too_small() {
+        let out = Hand::new(
+            vec![
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "Sw".to_string(),
+                "ShSo".to_string(),
+            ],
+            "3s".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidSuit);
+    }
+
+    #[test]
+    fn invalid_group_size_too_big() {
+        let out = Hand::new(
+            vec![
+                "SSSSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSw".to_string(),
+                "ShSo".to_string(),
+            ],
+            "3s".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidGroup);
+    }
+
+    #[test]
+    fn invalid_suit() {
+        let out = Hand::new(
+            vec![
+                "hhhjo".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSw".to_string(),
+            ],
+            "3s".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidSuit);
+    }
+
+    #[test]
+    fn hand_too_small() {
+        let out = Hand::new(
+            vec!["SSSw".to_string()],
+            "3s".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidShape);
+        let out = Hand::new(
+            vec!["SSSw".to_string()],
+            "3s".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidShape);
+    }
+    #[test]
+    fn hand_too_big() {
+        let out = Hand::new(
+            vec![
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSSw".to_string(),
+                "SSw".to_string(),
+                "SSSw".to_string(),
+            ],
+            "Sw".to_string(),
+            "3s".to_string(),
+            "3s".to_string(),
+        );
+        assert_eq!(out.unwrap_err(), HandErr::InvalidShape);
+    }
+}
